+++ src/R-xxhash-file.c
@@ -6,12 +6,12 @@
 
 #define XXH_STATIC_LINKING_ONLY   /* access advanced declarations */
 #include "xxhash.h"
-
+#include "R-xxhash-utils.h"
 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // XXH128
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-SEXP xxhash_file_xxh128(FILE *f) {
+SEXP xxhash_file_xxh128(FILE *f, SEXP as_raw_) {
   
   XXH3_state_t* state = XXH3_createState();
   XXH3_128bits_reset(state);
@@ -25,10 +25,7 @@ SEXP xxhash_file_xxh128(FILE *f) {
   XXH128_hash_t hash = XXH3_128bits_digest(state);
   XXH3_freeState(state);
   
-  char chash[32+1];
-  snprintf(chash, sizeof(chash), "%016" PRIx64 "%016" PRIx64, hash.high64, hash.low64);
-  
-  return mkString(chash);
+  return xxh128_hash_to_robj(hash, as_raw_);
 }
 
 
@@ -36,7 +33,7 @@ SEXP xxhash_file_xxh128(FILE *f) {
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // XXH3 64bit
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-SEXP xxhash_file_xxh3(FILE *f) {
+SEXP xxhash_file_xxh3(FILE *f, SEXP as_raw_) {
   
   XXH3_state_t* state = XXH3_createState();
   XXH3_64bits_reset(state);
@@ -50,10 +47,7 @@ SEXP xxhash_file_xxh3(FILE *f) {
   XXH64_hash_t hash = XXH3_64bits_digest(state);
   XXH3_freeState(state);
   
-  char chash[32+1];
-  snprintf(chash, sizeof(chash), "%016" PRIx64, hash);
-  
-  return mkString(chash);
+  return xxh64_hash_to_robj(hash, as_raw_);
 }
 
 
@@ -61,7 +55,7 @@ SEXP xxhash_file_xxh3(FILE *f) {
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // XXH32
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-SEXP xxhash_file_xxh32(FILE *f) {
+SEXP xxhash_file_xxh32(FILE *f, SEXP as_raw_) {
   
   XXH32_state_t *state = XXH32_createState();
   XXH32_reset(state, 0);
@@ -76,17 +70,14 @@ SEXP xxhash_file_xxh32(FILE *f) {
   XXH32_hash_t hash = XXH32_digest(state);
   XXH32_freeState(state);
   
-  char chash[32+1];
-  snprintf(chash, sizeof(chash), "%08x", hash);
-  
-  return mkString(chash);
+  return xxh32_hash_to_robj(hash, as_raw_);
 }
 
 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // XXH64
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-SEXP xxhash_file_xxh64(FILE *f) {
+SEXP xxhash_file_xxh64(FILE *f, SEXP as_raw_) {
   
   XXH64_state_t *state = XXH64_createState();
   XXH64_reset(state, 0);
@@ -101,10 +92,7 @@ SEXP xxhash_file_xxh64(FILE *f) {
   XXH64_hash_t hash = XXH64_digest(state);
   XXH64_freeState(state);
   
-  char chash[32+1];
-  snprintf(chash, sizeof(chash), "%016" PRIx64, hash);
-  
-  return mkString(chash);
+  return xxh64_hash_to_robj(hash, as_raw_);
 }
 
 
@@ -113,7 +101,7 @@ SEXP xxhash_file_xxh64(FILE *f) {
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // File
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-SEXP xxhash_file_(SEXP file_, SEXP algo_) {
+SEXP xxhash_file_(SEXP file_, SEXP algo_, SEXP as_raw_) {
   
   const char *file = CHAR(STRING_ELT(file_, 0));
   FILE *f = fopen(file, "rb");
@@ -124,13 +112,13 @@ SEXP xxhash_file_(SEXP file_, SEXP algo_) {
   const char *algo = CHAR(STRING_ELT(algo_, 0));
   SEXP res_ = R_NilValue;
   if (strcmp(algo, "xxh128") == 0) {
-    res_ = PROTECT(xxhash_file_xxh128(f));
+    res_ = PROTECT(xxhash_file_xxh128(f, as_raw_));
   } else if (strcmp(algo, "xxh3") == 0) {
-    res_ = PROTECT(xxhash_file_xxh3(f));
+    res_ = PROTECT(xxhash_file_xxh3(f, as_raw_));
   } else if (strcmp(algo, "xxh32") == 0) {
-    res_ = PROTECT(xxhash_file_xxh32(f));
+    res_ = PROTECT(xxhash_file_xxh32(f, as_raw_));
   } else if (strcmp(algo, "xxh64") == 0) {
-    res_ = PROTECT(xxhash_file_xxh64(f));
+    res_ = PROTECT(xxhash_file_xxh64(f, as_raw_));
   } 
   
   fclose(f);