+++ src/R-xxhash-utils.h
@@ -1,4 +1,4 @@
-SEXP xxh128_hash_to_robj(XXH128_hash_t hash, SEXP raw_);
-SEXP xxh64_hash_to_robj(XXH64_hash_t hash, SEXP raw_);
-SEXP xxh32_hash_to_robj(XXH32_hash_t hash, SEXP raw_);
+SEXP xxh128_hash_to_robj(XXH128_hash_t hash, SEXP as_raw_);
+SEXP xxh64_hash_to_robj(XXH64_hash_t hash, SEXP as_raw_);
+SEXP xxh32_hash_to_robj(XXH32_hash_t hash, SEXP as_raw_);