+++ src/init.c
@@ -2,18 +2,18 @@
 #include <R.h>
 #include <Rinternals.h>
 
-SEXP xxhash_(SEXP robj_, SEXP algo_);
-SEXP xxhash_raw_(SEXP robj_, SEXP algo_);
-SEXP xxhash_file_(SEXP file_, SEXP algo_);
+SEXP xxhash_     (SEXP robj_, SEXP algo_, SEXP as_raw_);
+SEXP xxhash_raw_ (SEXP robj_, SEXP algo_, SEXP as_raw_);
+SEXP xxhash_file_(SEXP file_, SEXP algo_, SEXP as_raw_);
 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // .C      R_CMethodDef
 // .Call   R_CallMethodDef
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 static const R_CallMethodDef CEntries[] = {
-  {"xxhash_"     , (DL_FUNC) &xxhash_     , 2},
-  {"xxhash_raw_" , (DL_FUNC) &xxhash_raw_ , 2},
-  {"xxhash_file_", (DL_FUNC) &xxhash_file_, 2},
+  {"xxhash_"     , (DL_FUNC) &xxhash_     , 3},
+  {"xxhash_raw_" , (DL_FUNC) &xxhash_raw_ , 3},
+  {"xxhash_file_", (DL_FUNC) &xxhash_file_, 3},
   {NULL, NULL, 0}
 };